generic list + table of source/shastra results
