--- shastra_compedium/templates/shastra_compedium/source_list.html
@@ -1,12 +1,0 @@
-{% extends 'base.html' %}
-{% block title %}Source Texts{% endblock %}
-{% block content %}
-<h1>Sources</h1>
-<ul>
-{% for source in object_list %}
-    <li>{{ source.title }}</li>
-{% empty %}
-    <li>No sources yet.</li>
-{% endfor %}
-</ul>
-{% endblock %}
